- Fixed up a couple enum issues that make complains about
git-svn-id: svn://70.113.44.214/VanaMirror/trunk@1916 09c1253f-3d09-8e40-9719-4320e56601b5

+++ src/ChannelServer/MobsPacket.cpp
@@ -134,7 +134,7 @@ void MobsPacket::damageMob(Player *player, PacketReader &pack) {
 				case 12: masteryid = Jobs::Page::SwordMastery; break;
 				case 90:
 				case 91:
-					masteryid = (player->getSkills()->getSkillLevel(Jobs::Fighter::SwordMastery) >= player->getSkills()->getSkillLevel(Jobs::Page::SwordMastery) ? Jobs::Fighter::SwordMastery : Jobs::Page::SwordMastery);
+					masteryid = (player->getSkills()->getSkillLevel(Jobs::Fighter::SwordMastery) >= player->getSkills()->getSkillLevel(Jobs::Page::SwordMastery) ? (int32_t)Jobs::Fighter::SwordMastery : (int32_t)Jobs::Page::SwordMastery);
 					break;
 			}
 			break;
